+++ client/src/components/Main.scss
@@ -1,12 +1,16 @@
 @use '../styles/constants' as *;
 
 main {
+  padding: 0 100px 70px;
+  word-break: break-all;
   min-height: calc(100vh - 443px);
   margin-top: 120px;
 }
 
 @media screen and (max-width: $md-breakpoint) {
   main {
+    padding: 0 20px 30px;
+
     min-height: calc(100vh - 718px);
     margin-top: 50px;
   }
